+++ OTAnalytics/plugin_ui/customtkinter_gui/dummy_viewmodel.py
@@ -1,10 +1,8 @@
 import contextlib
 from datetime import datetime
 from pathlib import Path
-from tkinter.filedialog import askopenfilename, askopenfilenames, asksaveasfilename
+from tkinter.filedialog import askopenfilename, askopenfilenames
 from typing import Iterable, Optional
-
-from adapter_ui.helpers import ensure_file_extension_is_present
 
 from OTAnalytics.adapter_ui.abstract_canvas import AbstractCanvas
 from OTAnalytics.adapter_ui.abstract_frame_canvas import AbstractFrameCanvas
@@ -55,6 +53,7 @@ from OTAnalytics.domain.section import (
 from OTAnalytics.domain.track import TrackId, TrackImage, TrackListObserver
 from OTAnalytics.domain.types import EventType
 from OTAnalytics.domain.video import Video, VideoListObserver
+from OTAnalytics.plugin_ui.customtkinter_gui.helpers import ask_for_save_file_path
 from OTAnalytics.plugin_ui.customtkinter_gui.line_section import (
     ArrowPainter,
     CanvasElementDeleter,
@@ -110,27 +109,6 @@ class MissingInjectedInstanceError(Exception):
 
 def flow_id(from_section: str, to_section: str) -> str:
     return f"{from_section} -> {to_section}"
-
-
-def ask_for_save_file_path(
-    title: str, filetypes: list[tuple[str, str]], defaultextension: str
-) -> Path:
-    """
-    Ask for a filename and ensure the file contains a file extension. If no extension
-    is present, the default extension will be appended.
-
-    Args:
-        title (str): title for the file chooser
-        file_types (list[tuple[str, str]]): supported file types to choose from
-        defaultextension (str): default extension used if none is present
-
-    Returns:
-        Path: path object representing an output path
-    """
-    filename = asksaveasfilename(
-        title=title, filetypes=filetypes, defaultextension=defaultextension
-    )
-    return ensure_file_extension_is_present(filename, defaultextension)
 
 
 class DummyViewModel(
